swinggrid/requirements.txt
# AutoSwing - Crypto Grid Trading Dashboard
# Python dependencies

streamlit>=1.28.0
python-binance>=1.0.19
plotly>=5.18.0
pandas>=2.0.0
python-decouple>=3.8
requests>=2.31.0
numpy>=1.24.0
